Rename Sha2 classes to SHA*
