+++ library/sha2/src/commonMain/kotlin/org/kotlincrypto/hash/SHA224.kt
@@ -1,0 +1,74 @@
+/*
+ * Copyright (c) 2023 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+package org.kotlincrypto.hash
+
+import org.kotlincrypto.core.Digest
+import org.kotlincrypto.core.internal.DigestState
+
+/**
+ * SHA-224 implementation
+ * */
+public class SHA224: Bit32Digest {
+
+    public constructor(): super(
+        d = 224,
+        h0 = -1056596264,
+        h1 = 914150663,
+        h2 = 812702999,
+        h3 = -150054599,
+        h4 = -4191439,
+        h5 = 1750603025,
+        h6 = 1694076839,
+        h7 = -1090891868,
+    )
+
+    private constructor(state: DigestState, sha224: SHA224): super(state, sha224)
+
+    protected override fun copy(state: DigestState): Digest = SHA224(state, this)
+
+    protected override fun out(a: Int, b: Int, c: Int, d: Int, e: Int, f: Int, g: Int, h: Int): ByteArray {
+        return byteArrayOf(
+            (a shr 24).toByte(),
+            (a shr 16).toByte(),
+            (a shr  8).toByte(),
+            (a       ).toByte(),
+            (b shr 24).toByte(),
+            (b shr 16).toByte(),
+            (b shr  8).toByte(),
+            (b       ).toByte(),
+            (c shr 24).toByte(),
+            (c shr 16).toByte(),
+            (c shr  8).toByte(),
+            (c       ).toByte(),
+            (d shr 24).toByte(),
+            (d shr 16).toByte(),
+            (d shr  8).toByte(),
+            (d       ).toByte(),
+            (e shr 24).toByte(),
+            (e shr 16).toByte(),
+            (e shr  8).toByte(),
+            (e       ).toByte(),
+            (f shr 24).toByte(),
+            (f shr 16).toByte(),
+            (f shr  8).toByte(),
+            (f       ).toByte(),
+            (g shr 24).toByte(),
+            (g shr 16).toByte(),
+            (g shr  8).toByte(),
+            (g       ).toByte(),
+        )
+    }
+}
